==> 03 Scripts/1. IC Data import and descriptive analysis.ipynb <==
List of contents
1. Import libraries
2. Import first data set
3. Explore first data set: descriptive statistics, column names, data types, create vars list
4. Import second data set
5. Explore second data set: descriptive statistics, column names, data types

1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Import orders data

# Import data (orders.csv) and create dataframe
df = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# Create string of path
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

3. Explore orders data

# Display table (first 5 rows)
df.head()

# Display table (top 30 rows)
df.head(30)

# Display end of table
df.tail()

# Display # of rows and columns
df.shape

# Display table details
df.info()

# Display data types
df.dtypes

# Display names of columns
df.columns

# Display descriptive statistics
df.describe()

# Create vars_list to include only necessary columns
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

# Select columns to display
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# Display new table (first 5 rows)
df.head()

4. Import second data set (products) 

# Import data (products.csv) and create dataframe
df_prods = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

# Create string of path (products.csv)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

5. Explore second data set (products)

# Display first 20 rows of table
df_prods.head(20)

# Display last 35 rows of table
df_prods.tail(35)

# Display names of columns
df_prods.columns

# Display numbers of rows and columns
df_prods.shape

# Display descriptive statistics
df_prods.describe()

# Display data types
df_prods.dtypes
