==> 03 Scripts/2. IC Data wrangling.ipynb <==
List of contents:
1. Import libraries
2. Import data sets
3. Data wrangling: dropping columns, renaming columns, changing data types, transposing data, creating data dictionaries, creating subsets
4. Export data sets

1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Importing data sets

# Importing orders data and create dataframe
df_ords = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# Create string of path (orders)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function (orders)
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Importing products data and create dataframe
df_prods = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

# Create string of path (products)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function (products)
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Importing departments data and create dataframe
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

3. Data wrangling

# Looking for missing values in 'days_since_prior_order' column in orders set
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Change order_dow column name in orders set
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# View orders data
df_ords.head()

# Change data type order_id column in orders set
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# View department data
df_dep.head()

# Transposing department dataframe
df_dep.T

# Overwrite old df_dep dataframe
df_dep_t = df_dep.T

# View new dataframe
df_dep_t

# Insert index df_dep
df_dep_t.reset_index()

# Remove first row of dp_dep_t for header
new_header = df_dep_t.iloc[0]

new_header

# Take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Set the header row as the df_header
df_dep_t_new.columns = new_header

df_dep_t_new

# Turn df_dep_t_new into a dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

# See first 5 entries for products set
df_prods.head()

# Determine what row 0 and department_id 19 indicate
print(data_dict.get('19'))

# Subsetting data in df_prods from the snack department
df_prods['department_id']==19

# Run indexing code again wrapped with another instance of df_prods[]
df_prods[df_prods['department_id'] == 19]

# Save results into new datafram df_snacks
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# Change user_id variable in orders set from numeric to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# Change column name without overwriting data frame
df_ords.rename(columns = {'order_hour_of_day' : 'hour_number_of_order'}, inplace = False)

# Check frequency of busiest hour for placing orders - Answer is 10:00
df_ords['order_hour_of_day'].value_counts()

# Determine meaning of department_value of 4 within the df_prods using data dictionary - Answer is produce
data_dict = df_dep_t_new.to_dict('index')

data_dict

# View department name
print(data_dict.get('4'))

# Create subset containing information about breakfast sales 
df_prods['department_id'] == 14

# View subset
df_prods[df_prods['department_id'] == 14]

# Create  new dataframe from subset
df_breakfast = df_prods[df_prods['department_id'] == 14]

# View dataframe
df_breakfast.head()

# Create subset for dinner parties including the following departments: alcohol, deli, beverages, meat/seafood
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

# View new dataframe
df_dinner_parties.head()

# Count number of rows in last dataframe created (df_dinner_parties) - Answer is 7,650
df_dinner_parties.shape

# Extract all information about user_id 1 - Answer is there appears to be no data for user id 1
df_user_id_1 = df_ords.loc[df_ords['user_id'] == 1]

# View dataframe
df_user_id_1.head()

# Provide basic stats on user_id 1 - Answer is that the descriptive statistics confirm that user_id 1 has missing values across all the columns in the orders set
df_user_id_1.describe()

4. Export data

# 12. Export df_ords datafram as 'order_wrangled.csv' in 'Prepared Data' folder
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# 13. Export df_dep_t_new dataframe as 'departments_wrangled.csv' in 'Prepared Data' folder
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments.wrangled.csv'))
