==> 03 Scripts/3 IC Data consistency check.ipynb <==
List of contents:
1. Import libraries
2. Import data
3. Data consistency checks: check for mixed-type data, missing values, and duplicates
4. Export data

1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Importing data

# Importing products data
df_prods = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

# Create string of path (products)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function (products)
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Importing orders data
df_ords = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_wrangled.csv', index_col = False)

# Create string of path (orders)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function (orders)
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

3. Data Consistency Checks

# Check descriptive statistics
df_ords.describe()

# Look for missing values
df_prods.isnull().sum()

# View missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Find number of rows
df_prods.shape

# Create new dataframe filtering out missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Recheck number of rows
df_prods_clean.shape

# Look for full duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Check number of rows
df_prods_clean.shape

# Create new dataframe with no duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# Check number of rows in new dataframe
df_prods_clean_no_dups.shape

# Check descriptive statistics of df_prods
df_prods.describe()

# Check for mixed-type data in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# Check for missing values in df_ords
df_ords.isnull().sum()

# View missing values in df_ords
df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_nan

# Find number of rows in df_ords
df_ords.shape

# Check for duplicates in df_ords
df_dups = df_ords[df_ords.duplicated()]

df_dups

4. Export data

# Export df_ords as a data file in Prepared Data folder with new orders_checked name
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))

# Export df_prods_clean_no_dups dataframe as 'products_checked.csv' in 'Prepared Data' folder
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))
