==> 03 Scripts/4.1. IC Combining and exporting data.ipynb <==
List of contents:
1. Importing libraries
2. Importing data
3. Checking data - check output and number of rows
4. Merge data sets together into new dataframe
5. Export new dataframe as pkl

1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Importing data

# Importing orders_product_prior data
df_ords_prior = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\order_products__prior.csv', index_col = False)

# Create string of path
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function 
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'order_products__prior.csv'), index_col = False)

# Importing orders_checked data
df_ords = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_checked.csv', index_col = False)

# os path function 
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

 3. Checking data

# Check the output
df_ords_prior.head()

# Check the output
df_ords.head()

# Check number of rows
df_ords_prior.shape

# Check number of rows
df_ords.shape

4. Create new dataframe with merge of both data sets

# Create new dataframe
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the output
df_merged_large.head()

# Confirm merge results
df_merged_large['_merge'].value_counts()

5. Export new dataframe

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
