==> 03 Scripts/4.2 IC Combining and exporting data.ipynb <==
List of contents:
1. Import libraries
2. Import data
3. Checking data consistency - check output and number of rows and columns of data sets
4. Merge data sets into new dataframe
5. Export new dataframe as pkl

1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Import data

# Import order_products_combined pkl
df_orders_products_combined = pd.read_pickle(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')

# Create string of path
path = r'C:\Users\court\Instacart Basket Analysis'

path

# Import products data
df_products = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\products_checked.csv', index_col = False)

3. Checking data

# Check shape of pkl file
df_orders_products_combined.shape

# Check shape of products_checked data
df_products.shape

# Check the output of pkl file
df_orders_products_combined.head()

# Check the output of products data
df_products.head()

df_products_checked = df_products.drop(columns = ['Unnamed: 0'])

df_products_checked.head()

df_orders_products_combined = df_orders_products_combined.drop(columns = ['Unnamed: 0'])

df_orders_products_combined = df_orders_products_combined.drop(columns = ['Unnamed: 0.1'])

df_orders_products_combined.head()

4. Merge data

# Create a new dataframe
df_large_merged_2 = df_orders_products_combined.merge(df_products_checked, on = 'product_id')

# Confirm results of merge
df_large_merged_2['_merge'].value_counts(dropna=False)

df_products_checked.merge(df_orders_products_combined, on = ['product_id'], how = 'outer')

5. Export new dataframe

# Export data to pkl
df_large_merged_2.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
