==> 03 Scripts/5. IC Data immersion.ipynb <==
List of contents:
1. Import libraries
2. Import data
3. Derive new variables: user-defined functions, loc(), with for-loops
4. Export data as pkl

1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Import data

# Import order_products_combined pkl
df_orders_products_merged = pd.read_pickle(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')

# Create string of path
path = r'C:\Users\court\Instacart Basket Analysis'

path

3. Derive new variables

# Create subset of data
df = df_orders_products_merged[:1000000]

df.shape

# Define a function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function
df['price_range'] = df.apply(price_label, axis=1)

# Check frequency
df['price_range'].value_counts(dropna = False)

# Check maximum value
df['prices'].max()

# Create function with loc() for price_range
df_orders_products_merged.loc[df_orders_products_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_orders_products_merged.loc[(df_orders_products_merged['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_orders_products_merged.loc[df_orders_products_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check frequency
df_orders_products_merged['price_range_loc'].value_counts(dropna = False)

# Check frequency
df_orders_products_merged['orders_day_of_week'].value_counts(dropna = False)

# Create a new column with for loops for order_day_of_week
result = []

for value in df_orders_products_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# View results list
df_orders_products_merged['busiest_day'] = result

# Check frequency
df_orders_products_merged['busiest_day'].value_counts(dropna = False)

# Create busiest day column containing top 2 busiest and slowest days
df_orders_products_merged.loc[df_orders_products_merged['orders_day_of_week'] <= 1, 'busiest_day_loc'] = 'Busiest days'

df_orders_products_merged.loc[(df_orders_products_merged['orders_day_of_week'] > 2) & (df['orders_day_of_week'] < 5), 'busiest_day_loc'] = 'Slowest days'

# Check frequency
df_orders_products_merged['busiest_day_loc'].value_counts(dropna = False)

# Check values regarding to time of day for order
df_orders_products_merged['order_hour_of_day'].value_counts(dropna = False)

# Create column - busiest_period_of_day
def busiest_period_of_day(row):

  if row['order_hour_of_day'] <= 5:
    return 'Fewest orders'
  elif (row['order_hour_of_day'] > 9) and (row['order_hour_of_day'] <= 16):
    return 'Most orders'
  else: return 'Average orders'

# Apply the function
df['order_hour_of_day'] = df.apply(busiest_period_of_day, axis=1)

# Check frequency in new column
df['order_hour_of_day'].value_counts(dropna = False)

4. Export data as pkl file

# Import as pkl file
df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
