==> 03 Scripts/6. IC Grouping data and aggregating variables.ipynb <==
List of contents:
1. Importing libraries
2. Importing data
3. Aggregating data: grouping with pandas, with agg(), with transform(), deriving with loc()
4. Exporting data as pkl

1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Importing data

# Import order_products_immersed pkl
df_ords_prods_merge = pd.read_pickle(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')

# Create string of path
path = r'C:\Users\court\Instacart Basket Analysis'

path

3. Aggregating data

# Create a subset
df = df_ords_prods_merge[:1000000]

# Aggregate mean of order_number grouped by department_id for entire dataframe
df_ords_prods_merge.groupby('department_id').agg({'order_number' : ['mean']})

# Group data
df['max_order'] = df.groupby(['user_id'])['order_number'].transform(np.max)

# View data with new column
df.head(10)

# Aggregate data using loyalty flag
df.loc[df['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df.loc[(df['max_order'] <= 40) & (df['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

df.loc[df['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Display frequency counts
df['loyalty_flag'].value_counts(dropna = False)

# View data by certain columns
df[['user_id', 'loyalty_flag', 'prices', 'add_to_cart_order']].head(60)

# Group data
df['mean_price'] = df.groupby(['user_id'])['prices'].transform(np.mean)

# View data with new column
df.head(10)

# Create a spending flag
df.loc[df['mean_price'] < 10, 'spending_flag'] = 'Low spender'

df.loc[df['mean_price'] >= 10, 'spending_flag'] = 'High spender'

# View frequency counts
df['spending_flag'].value_counts(dropna = False)

# View data with certain columns
df[['user_id', 'spending_flag', 'prices']].head(60)

# Group data
df['median_days'] = df.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# View data with new column
df.head(10)

# Aggregate data with frequency flag
df.loc[df['median_days'] > 20, 'frequency_flag'] = 'Non-frequent customer'

df.loc[(df['median_days'] <= 20) & (df['median_days'] > 10), 'frequency_flag'] = 'Regular customer'

df.loc[df['median_days'] <= 10, 'frequency_flag'] = 'Frequent customer'

# View frequency counts
df['frequency_flag'].value_counts(dropna = False)

# View data with certain columns
df[['user_id', 'frequency_flag', 'order_id']].head(60)

4. Export file as pkl

# Export df at pkl file
df.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
