==> 03 Scripts/7.1. IC Data visualization.ipynb <==
List of contents:
1. Import libraries
2. Import data
3. Data quality checks: output, number of rows, column names, data details, data types, descriptive statistics, vars list
4. Data wrangling: change column names and data types
5. Data consistency missing values and duplicates
6. Merge data and create new dataframe
7. Import dataframe as pkl

1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

2. Import data

# Importing customers data
df_custs = pd.read_csv(r'C:\Users\court\Instacart Basket Analysis\02 Data\Original Data\customers.csv', index_col = False)

# Create string of path (orders)
path = r'C:\Users\court\Instacart Basket Analysis'

path

# os path function (customers)
df_custs = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

# Importing ords_prods_merge data
df_orders_products_final = pd.read_pickle(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')

3. Check data quality

# Display output
df_custs.head()

# Display last 5 rows
df_custs.tail()

# Check number of rows and columns
df_custs.shape

# Display table details
df_custs.info()

# Display data types
df_custs.dtypes

# Display names of columns
df_custs.columns

# Display descriptive statistics
df_custs.describe()

# Create vars list
vars_list = ['user_id', 'First Name', 'Surnam', 'Gender', 'STATE', 'Age',
       'date_joined', 'n_dependants', 'fam_status', 'income']

vars_list

# Select columns to display
df_custs = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), usecols = vars_list)

# Display new table
df_custs.head()

# Check shape of pkl file
df_orders_products_final.shape

# Check shape of customers data
df_custs.shape

# Check the output of pkl file
df_orders_products_final.head()

4. Data wrangling

# Change column names
df_custs.rename(columns = {'First Name' : 'first_name'}, inplace = True)

df_custs.rename(columns = {'Surnam' : 'last_name'}, inplace = True)

df_custs.rename(columns = {'Gender' : 'gender'}, inplace = True)

df_custs.rename(columns = {'STATE' : 'state'}, inplace = True)

df_custs.rename(columns = {'Age' : 'age'}, inplace = True)

df_custs.rename(columns = {'n_dependants' : 'number_of_dependants'}, inplace = True)

df_custs.rename(columns = {'fam_status' : 'family_status'}, inplace = True)

# Check new column names
df_custs.head()

# Change data type
df_custs['user_id'] = df_custs['user_id'].astype('str')

# Check new data type
df_custs['user_id'].dtype

5. Data consistency

# Look for missing values
df_custs.isnull().sum()

# View missing values
df_nan = df_custs[df_custs['first_name'].isnull() == True]

df_nan

df_nan.shape

# Look for full duplicates
df_dups = df_custs[df_custs.duplicated()]

df_dups

# Change data type for merge
df_orders_products_final['user_id'] = df_orders_products_final['user_id'].astype('str')

# Check new data type
df_orders_products_final['user_id'].dtype

6. Merge data

# Create a new dataframe
df_large_merged = df_custs.merge(df_orders_products_final, on = 'user_id', how = 'inner')

# Confirm results of merge
df_large_merged['_merge'].value_counts()

# View new dataframe
df_large_merged.head()

7. Export data as pkl file

# Export data to pkl
df_large_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
