==> 03 Scripts/8.1 IC Last steps.ipynb <==
List of contents:
1. Import libraries
2. Import pkl file
3. Security implications check: remove PII
4. Exclusion flag: data sample, export of sample
5. Data derivation
6. Export pkl file

1. Import libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

2. Import pkl file

df_ords_prods_custs = pd.read_pickle(r'C:\Users\court\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_all.pkl')

# Create path
path = r'C:\Users\court\Instacart Basket Analysis'

path

3. Check for security implications

# Remove columns containing PII data
df_final = df_ords_prods_custs.drop(columns = ['first_name', 'last_name'])

#View data for PII data
df_final.head()

4. Exclusion flag

# Mark customers with less than 5 orders as missing values
df_final.loc[df_final['max_order'] <5, 'max_order'] = np.nan

# Check newest minimal order amount
df_final['max_order'].min()

# View missing values
df_nan = df_final[df_final['max_order'].isnull() == True]

df_nan

# Count how many 'max_orders' having missing values
df_nan['max_order'].value_counts(dropna = False)

# Find number of rows in original df
df_final.shape

# Create new dataframe filtering out missing values
df_final_clean = df_final[df_final['max_order'].isnull() == False]

# Check number of rows in df
df_final_clean.shape

# Make sample of df_final_clean
np.random.seed(4)

dev = np.random.rand(len(df_final_clean)) <=0.7

# Make big df with 70% of data
big = df_final_clean[dev]

# Make small df with 30% of data
small = df_final_clean[~dev]

# Check number of rows and columns 
len(df_final_clean)

len(big) + len(small)

# Export sample of df_final_clean as pkl
small.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'data_without_low_activity_customers.pkl'))

5. Data derivation

# Assign states to regions
df_final.loc[df_final['state'].isin(['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']), 'region'] = 'Northeast'

df_final.loc[df_final['state'].isin(['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']), 'region'] = 'Midwest'

df_final.loc[df_final['state'].isin(['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Lousiana']), 'region'] = 'South'

df_final.loc[df_final['state'].isin(['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']), 'region'] = 'West'

# View frequency counts
df_final['region'].value_counts(dropna = False)

# Crosstab to further analyze data of regions and spending habits
crosstab = pd.crosstab(df_final['region'], df_final['spending_flag'], dropna = False)

crosstab.to_clipboard()

# Assign flags to different demographics

def customer_age(row):

  if row['age'] <= 39:
    return 'Young adults'
  elif (row['age'] > 39) and (row['age'] <= 64):
    return 'Middle age adults'
  elif row['age'] > 64:
    return 'Older adults'
  else: return 'Not enough data'

# Apply the function
small['age_group'] = small.apply(customer_age, axis=1)

# Check frequency
small['age_group'].value_counts(dropna = False)

# Define function for multiple customer characteristics

def characteristics_label(row):

  if (row['number_of_dependants'] <= 0) and (row['family_status'] == 'single'):
    return 'Single adult'
  elif (row['number_of_dependants'] == 1) and (row['family_status'] == 'married'):
    return 'Couple with no kids'
  elif (row['number_of_dependants'] <= 0) and (row['family_status'] == 'divorced/widowed'):
    return 'Other single adult'
  elif (row['number_of_dependants'] <= 0) and (row['family_status'] == 'living with parents and siblings'):
    return 'Single adult still living with parents'
  elif (row['number_of_dependants'] > 1) and (row['age'] > 40):
    return 'Older parents'
  elif (row['number_of_dependants'] > 1) and (row['age'] < 40):
    return 'Younger parents'
  else: return 'Other'

# Apply the function
small['unique_customers'] = small.apply(characteristics_label, axis=1)

# Check frequency
small['unique_customers'].value_counts(dropna = False)

6. Export pkl file

# Export small as pkl for further analysis
small.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))
